+++ Orchestration/PanelSourceOrchestrator.cs
@@ -44,8 +44,7 @@ namespace MSFSPopoutPanelManager.Orchestration
         {
             if (ActiveProfile.IsSelectingPanelSource)
                 return;
-
-            WindowProcessManager.SetApplicationProcess();
+                        
             ActiveProfile.IsSelectingPanelSource = true;
         }