+++ RELEASENOTES.md
@@ -1,4 +1,5 @@
 ## Version 1.0.3
 
-* Added - ChasePlane 2024 support
-* Added - Option to close Pop Out Panel Manager after successfully executing pop out process
+* Added - ChasePlane 2024 support.
+* Added - Option to close Pop Out Panel Manager after successfully executing pop out process.
+* Fixed - Application crash when executable file was renamed.